[int][new][bt] [modify avrcp get element function]
Project: Bluetooth

redmine: #4703, REDMINE-4703

ext-redmine: bug|feat#4703

[Description in detail]

Affected branch: [master]

+++ service/bt/bt_finsh/bts2_app_interface_type.h
@@ -300,7 +300,7 @@ typedef struct
 typedef struct
 {
     uint8_t  track_id;
-    uint8_t  attri_req;
+    uint32_t  attri_req;
     bt_notify_avrcp_music_detail_info_t  detail_info;
 } bt_notify_avrcp_music_detail_t;
 #endif
